[DSD] Correctly handle shared parameters for optimizer state_dict
Pull Request resolved: https://github.com/pytorch/pytorch/pull/128685

Current implementation of `set_optimizer_state_dict()` assumes that all the fqns returned by `_get_fqns()` must exist in the optimizer state_dict. This is not true if the model has shared parameters. In such a case, only one fqn of the shared parameters will appear in the optimizer state_dict. This PR addresses the issue.
ghstack-source-id: 230252126

Differential Revision: [D58573487](https://our.internmc.facebook.com/intern/diff/D58573487/)